Educational Forum
This is just a design for an online course websites
The concept is basically very simple. Here people who are willing to learn and gain new skills to their portfolio can sign up.
Lok for the desired course in the domain they want to gain opportunities and can learn the skill watching videos.
These video courses will be created by the online educators who will teach the specific skill to the students.